Added /encounter-method/ endpoint
