+++ src/api/components/encounter-method/encounter-method-repository.js
@@ -1,0 +1,28 @@
+const encounterMethod = require('../../../models')['encounter-method'];
+
+async function getByName(name) {
+    return encounterMethod.find(
+        { name },
+        {
+            '_id': 0,
+            '__v': 0,
+            'names._id': 0,
+        }
+    );
+}
+
+async function getByID(id) {
+    return encounterMethod.find(
+        { id },
+        {
+            '_id': 0,
+            '__v': 0,
+            'names._id': 0,
+        }
+    );
+}
+
+module.exports = {
+    getByID,
+    getByName,
+};
