+++ src/api/components/encounter-method/encounter-method-service.js
@@ -1,0 +1,12 @@
+const repository = require('./encounter-method-repository');
+
+async function getItem(name) {
+    if (/^-?\d+(\.\d+)?$/.test(name)) {
+        return repository.getByID(name);
+    }
+    return repository.getByName(name);
+}
+
+module.exports = {
+    getItem,
+};
