--- src/api/components/encounter-method/repository.js
@@ -1,14 +1,0 @@
-const { template } = require('../../../models');
-
-async function getTemplate() {
-    return template.find({});
-}
-
-async function create(body) {
-    return template.create(body);
-}
-
-module.exports = {
-    getTemplate,
-    create,
-};
